Initial conversion to Python 3
- automatic part done via 2to3
- several manual fixes to make initial page and concordance page work
  (not with any possible plug-ins)

+++ lib/plugins/default_conc_persistence/__init__.py
@@ -66,7 +66,7 @@ def mk_short_id(s, min_length):
     s -- a string to be hashed
     min_length -- minimum length of the output hash
     """
-    x = long('0x' + hashlib.md5(s).hexdigest(), 16)
+    x = int('0x' + hashlib.md5(s).hexdigest(), 16)
     ans = []
     while x > 0:
         p = x % len(KEY_ALPHABET)
